feat(esm): moving sendgrid data to esm
