+++ lambdas/sendgrid-data/package.json
@@ -7,6 +7,7 @@
     "dist",
     "package.json"
   ],
+  "type": "module",
   "scripts": {
     "build": "rm -rf dist && tsc",
     "format": "eslint --fix",
